+++ src/utils/logger.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const pino = require('pino')
-const app = require('../package')
+const app = require('../../package.json')
 
 module.exports = pino({
   name: app.name,
